Fix incorrect "Should print:" comment in ExampleStrictPolicy().
The "Should print:" comment in ExampleStrictPolicy() suggested the output should be "Hello  World", which is incorrect, as the actual output is "Goodbye  World".
Change examples to use concluding "Output:" line comment, which is then compared to actual standard output of the function when when the tests are run. That way, such a mistake would not occur without being noticed. See http://godoc.org/testing#hdr-Examples for reference.