Resolves issue #18
If an element was not permitted to allow no attributes, then we nuked it and would set a flag to remove the closing tag. This flag did not adequately account for nesting of tags and would omit the next closing tag it found... this was the source of the bug.

Whilst we continue to not support the balancing of a tree, and so in some edge cases deep nesting of the same element may still cause this bug, I have added a change in which the closing tag is identified and only the next closing tag that matches the omitted opening tag is the one that will be omitted.