Resolves #23
The bug was that if we skip element contents, but the contents themselves contain
another pair of tags that are also being skipped, the inner tags would terminate
the skipping of content early.

The fix is to add card-counting to ensure that however deep we are we do not
restore content until we have seen pairs of tags that bring us back to the top
level.