Add example command line utilities
Two utilities, one to demonstrate a user generated policy (the same config I use)
and the other to demonstrate normalising HTML email.

The HTML email example is not safe, and it not recommended, but if you are going
to do it then at least do it as safely as possible.