Resolves #51 by permitting spaces in URLs within HTML
It turns out that line feeds and white space within URLs within HTML is valid.
The HTML rules are that a URL is a valid URL if when trimmed of whitespace it
is a valid URL, or if whitespace is not considered it is a valid URL. Browsers
are fine with this, and the bluemonday behaviour of treating URLs as not being
part of HTML was incorrect in that we would reject URLs that had whitespace.

By permitting whitespace, we fix a known bug wherein linefeeds within data-uri
values would not be considered valid.