Updated for Go1.12
A change in net/url now fails to parse a URI if it contains control
chars like new lines. This is good, but out tests were written with nice
formatting and... new lines. Making the tests more realistic by removing
the new lines made the code work as expected.

Typically I do not like changing tests on version increments, but in
this case the test was wrong.