Add CSS styling for element attributes
This change adds in default handlers and functionality to add handlers
to specific styles. Building on top of Nicholls changes, this allows
for setting allowed styles as an attribute of either individual
elements or globally and validating using either a regular expression,
matching against a list of values, or with a supplied handler. If
none of these are supplied, it relies on a default handler for each
element. Please note this change does not allow for style elements,
or for external styule sheets.