go fmt with go 1.17
The build tags have changed... I am highly sceptical of what this means for older go versions whose tests will likely now fail.

But as these are build tags, I cannot see how I can avoid updating them to the newer format.

This is the way for the future... if it breaks the past then hmmmm.